// types/type.d.ts
import { TextInputProps, TouchableOpacityProps } from "react-native";

declare interface ButtonProps extends TouchableOpacityProps {
  title: string;
  bgVariant?: "primary" | "secondary" | "danger" | "outline" | "success";
  textVariant?: "primary" | "default" | "secondary" | "danger" | "success";
  IconLeft?: React.ComponentType<any>;
  IconRight?: React.ComponentType<any>;
  className?: string;
}

declare interface InputFieldProps extends TextInputProps {
  title: string;
  icon?: any;
  secureTextEntry?: boolean;
  labelStyle?: string;
  containerContext?: string;
  inputStyle?: string;
  iconStyle?: string;
  className?: string;
  LeftIcone?: React.ComponentElement<any>;
  RightIcone?: React.ComponentElement<any>;
}